# contracts/openapi/pramaan-backend.yaml
openapi: 3.1.0
info:
  title: Pramaan Next — Backend (Phase 1)
  version: 0.1.0
  description: >
    Evidence ingestion, versioning, integrity, deterministic repository queries,
    application metadata, and the collection scheduler. Phase 1 uses deterministic
    mock enterprise integrations. No LLM. Source of truth for backend ⇄ frontend
    and backend ⇄ agents interfaces — implement against this file, do not hand-roll.
servers:
  - url: http://localhost:8080
    description: Local dev
...
paths:
  /api/v1/insight/completeness:
    get:
      tags: [insight]
      summary: Expected-control coverage for one application (deterministic)
      parameters:
        - { name: applicationSlug, in: query, required: true, schema: { type: string } }
        - { name: framework, in: query, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CompletenessReport" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/insight/compliance:
    get:
      tags: [insight]
      summary: Compliance posture rollup (check verdicts × expected controls, deterministic)
      parameters:
        - { name: applicationSlug, in: query, required: true, schema: { type: string } }
        - { name: framework, in: query, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ComplianceReport" } } } }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/insight/comparison:
    get:
      tags: [insight]
      summary: "UC14 — cross-application compliance comparison (matrix + gap list)"
      parameters:
        - { name: applications, in: query, required: false, schema: { type: array, items: { type: string } } }
        - { name: framework, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ComparisonReport" } } } }
        "400": { $ref: "#/components/responses/Error" }
...
  /api/v1/insight/audit-prep:
    get:
      tags: [insight]
      summary: "UC18 — AI-assisted audit-preparation checklist (deterministic findings + grounded narrative)"
      parameters:
        - { name: applicationSlug, in: query, required: false, schema: { type: string } }
        - { name: framework, in: query, required: false, schema: { type: string } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/AuditPrepReport" } } } }
...
  /api/v1/reports/{name}:
    parameters: [ { name: name, in: path, required: true, schema: { type: string } } ]
    get:
      tags: [reports]
      summary: "Render a report as JSON (default) or CSV — a view over existing insight/evidence data"
      parameters:
        - { name: format, in: query, required: false, schema: { type: string, enum: [json, csv], default: json } }
        - { name: applicationSlug, in: query, required: false, schema: { type: string } }
        - { name: framework, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json: { schema: { type: object } }
            text/csv: { schema: { type: string } }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/insight/evidence/{id}/summary:
    parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [insight]
      summary: AI summary of one evidence record, grounded in stored content + check verdicts
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/EvidenceSummary" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/insight/reuse/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: limit, in: query, schema: { type: integer, default: 5 } }
      - { name: minScore, in: query, schema: { type: number, default: 0.3 } }
    get:
      tags: [insight]
      summary: Evidence similar to the given record (embedding nearest-neighbour + reuse hints)
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ReuseResult" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/insight/reuse/search:
    post:
      tags: [insight]
      summary: Evidence similar to a free-text query
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
                limit: { type: integer, default: 5 }
                minScore: { type: number, default: 0.3 }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ReuseResult" } } } }
  /api/v1/insight/nl-query:
    post:
      tags: [insight]
      summary: Natural-language audit query — deterministic routing for structured questions, RAG (embedding/pgvector retrieval, cited + grounded) for "what evidence do we have for X", AI narration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [question]
              properties:
                question: { type: string }
                applicationSlug: { type: string }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/NlQueryResult" } } } }
        "400": { $ref: "#/components/responses/Error" }
...
  /api/v1/controls:
    get:
      tags: [controls]
      summary: List controls (reference catalog)
      parameters:
        - { name: framework, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/ControlView" } }
  /api/v1/controls/{id}:
    parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
    get:
      tags: [controls]
      summary: Get one control
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/ControlView" } } } }
        "404": { $ref: "#/components/responses/Error" }
...
  /api/v1/checks:
    get:
      tags: [checks]
      summary: List predefined technical checks (collector catalog)
      parameters:
        - { name: collector, in: query, schema: { type: string } }
        - { name: framework, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/CheckDefView" } }
  /api/v1/checks/evaluate:
    post:
      tags: [checks]
      summary: Re-run deterministic rule evaluation over stored evidence (optionally narrowed)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                applicationSlug: { type: string }
                framework: { type: string }
                controlId: { type: string }
                sourceSystem: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  evaluatedAt: { type: string, format: date-time }
                  evidenceEvaluated: { type: integer }
                  resultsWritten: { type: integer }
                  byStatus: { type: object, additionalProperties: { type: integer } }
  /api/v1/check-results:
    get:
      tags: [checks]
      summary: Deterministic technical-check results (PASS / WARNING / FAIL / NOT_APPLICABLE)
      parameters:
        - { name: applicationSlug, in: query, schema: { type: string } }
        - { name: framework, in: query, schema: { type: string } }
        - { name: controlId, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: "#/components/schemas/CheckStatus" } }
        - { name: runId, in: query, schema: { type: string, format: uuid } }
        - { name: page, in: query, schema: { type: integer, default: 0 } }
        - { name: size, in: query, schema: { type: integer, default: 20, maximum: 500 } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      items: { type: array, items: { $ref: "#/components/schemas/CheckResultView" } }
...
  /api/v1/insight/evidence/{id}/context:
    parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [insight]
      summary: Source technology, collection method, controls satisfied and frameworks for one evidence record
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/EvidenceContext" } } } }
        "404": { $ref: "#/components/responses/Error" }

  # ---- Use Case 5 — ECS Admin (users, roles, applications) --------------
  # Applications are administered via /api/v1/applications (above) — not duplicated here.
...
components:
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }
  schemas:
    ErrorBody:
      type: object
      properties:
        timestamp: { type: string, format: date-time }
        status: { type: integer }
        error: { type: string }
        message: { type: string }
    Page:
      type: object
      properties:
        items: { type: array, items: {} }
        page: { type: integer }
        size: { type: integer }
        totalItems: { type: integer, format: int64 }
        totalPages: { type: integer }
...
    ControlView:
      type: object
      properties:
        id: { type: string }
        framework: { type: string }
        title: { type: string }
        domain: { type: string }
    CheckDefView:
      type: object
      description: One predefined technical check an agent collector can run.
      properties:
        checkId: { type: string }
        collector: { type: string, description: "os | database | middleware | tls" }
        technology: { type: string }
        controlId: { type: string }
        framework: { type: string }
        title: { type: string }
...
    CheckStatus:
      type: string
      enum: [PASS, WARNING, FAIL, NOT_APPLICABLE]
    CheckResultView:
      type: object
      properties:
        id: { type: string }
        runId: { type: string, format: uuid, nullable: true }
        agentId: { type: string, nullable: true }
        assetId: { type: string, nullable: true }
        applicationSlug: { type: string }
        checkId: { type: string }
        controlId: { type: string }
        framework: { type: string }
        status: { $ref: "#/components/schemas/CheckStatus" }
        observed: { type: string }
        expected: { type: string }
        detail: { type: string }
        evidenceId: { type: string, format: uuid, nullable: true }
        collectedAt: { type: string, format: date-time }
    ControlCoverage:
      type: object
      properties:
        framework: { type: string }
        controlId: { type: string }
        title: { type: string }
        coverage: { type: string, enum: [COVERED, STALE, MISSING] }
        evidenceId: { type: string, format: uuid, nullable: true }
        currentVersion: { type: integer, nullable: true }
        lastCollectedAt: { type: string, format: date-time, nullable: true }
        ageDays: { type: integer, nullable: true }
    CompletenessReport:
      type: object
      properties:
        applicationSlug: { type: string }
        framework: { type: string, nullable: true }
        generatedAt: { type: string, format: date-time }
        staleAfterDays: { type: integer }
        expected: { type: integer }
        covered: { type: integer }
        stale: { type: integer }
        missing: { type: integer }
        completenessPct: { type: number }
        controls: { type: array, items: { $ref: "#/components/schemas/ControlCoverage" } }
    SimilarEvidence:
      type: object
      properties:
        evidenceId: { type: string, format: uuid }
        applicationSlug: { type: string }
        framework: { type: string }
        controlId: { type: string }
        sha256: { type: string, nullable: true }
        score: { type: number }
        crossApplication: { type: boolean }
        sameControl: { type: boolean }
        exactDuplicate: { type: boolean, description: "byte-identical (same SHA-256) as the query evidence" }
        reuseHint: { type: string }
    ReuseResult:
      type: object
      properties:
        queryEvidenceId: { type: string, format: uuid, nullable: true }
        queryText: { type: string, nullable: true }
        embeddingModel: { type: string }
        vectorStore: { type: string, enum: [memory, pgvector] }
        indexed: { type: integer }
        querySha256: { type: string, nullable: true }
        exactDuplicates:
          type: array
          description: Other persisted records whose current version has the identical SHA-256.
          items: { $ref: "#/components/schemas/SimilarEvidence" }
        matches: { type: array, items: { $ref: "#/components/schemas/SimilarEvidence" } }
    EvidenceSummary:
      type: object
      properties:
        evidenceId: { type: string, format: uuid }
        applicationSlug: { type: string }
        framework: { type: string }
        controlId: { type: string }
        model: { type: string }
        simulated: { type: boolean }
        summary: { type: string }
        groundedOn: { type: array, items: { type: string } }
        generatedAt: { type: string, format: date-time }
    NlQueryResult:
      type: object
      properties:
        question: { type: string }
        interpretedAs: { type: string }
        matchedQuery: { type: string }
        answer: { type: object, additionalProperties: true }
        narrative: { type: string }
        model: { type: string }
        simulated: { type: boolean }
        generatedAt: { type: string, format: date-time }
    FrameworkPosture:
      type: object
      properties:
        framework: { type: string }
        expected: { type: integer }
        compliant: { type: integer }
        partiallyCompliant: { type: integer }
        nonCompliant: { type: integer }
        notAssessed: { type: integer }
        missingEvidence: { type: integer }
        compliancePct: { type: number }
    ControlPosture:
      type: object
      properties:
        framework: { type: string }
        controlId: { type: string }
        status:
          type: string
          enum: [COMPLIANT, PARTIALLY_COMPLIANT, NON_COMPLIANT, NOT_ASSESSED, MISSING_EVIDENCE]
        detail: { type: string }
    ComplianceReport:
      type: object
      properties:
        applicationSlug: { type: string }
        generatedAt: { type: string, format: date-time }
        expected: { type: integer }
        compliant: { type: integer }
        compliancePct: { type: number }
        byFramework: { type: array, items: { $ref: "#/components/schemas/FrameworkPosture" } }
        controls: { type: array, items: { $ref: "#/components/schemas/ControlPosture" } }
...
    ComparisonReport:
      type: object
      properties:
        generatedAt: { type: string, format: date-time }
        framework: { type: string, nullable: true }
        applications: { type: array, items: { type: string } }
        frameworks:
          type: array
          items:
            type: object
            properties:
              framework: { type: string }
              compliancePctByApp: { type: object, additionalProperties: { type: number } }
              minPct: { type: number }
              maxPct: { type: number }
              spreadPct: { type: number }
        controls:
          type: array
          items:
            type: object
            properties:
              framework: { type: string }
              controlId: { type: string }
              statusByApp: { type: object, additionalProperties: { type: string } }
              consistent: { type: boolean }
        gaps:
          type: array
          items:
            type: object
            properties:
              applicationSlug: { type: string }
              framework: { type: string }
              controlId: { type: string }
              status: { type: string }
...
    AuditPrepReport:
      type: object
      properties:
        generatedAt: { type: string, format: date-time }
        scope: { type: string }
        readinessScore: { type: number }
        totalFindings: { type: integer }
        bySeverity: { type: object, additionalProperties: { type: integer } }
        findings:
          type: array
          items:
            type: object
            properties:
              category: { type: string }
              severity: { type: string, enum: [HIGH, MEDIUM, LOW] }
              applicationSlug: { type: string }
              framework: { type: string }
              controlId: { type: string }
              detail: { type: string }
              action: { type: string }
        narrative: { type: string }
        model: { type: string }
        simulated: { type: boolean }
...
    EvidenceContext:
      type: object
      properties:
        evidenceId: { type: string, format: uuid }
        applicationSlug: { type: string }
        controlId: { type: string }
        technology: { type: string }
        collectionMethod: { type: string }
        sourceSystem: { type: string }
        collectedAt: { type: string, format: date-time, nullable: true }
        currentVersion: { type: integer }
        satisfiesControls: { type: array, items: { type: string } }
        frameworks: { type: array, items: { type: string } }